Use the qmlEngine method instead of creating a new QQmlEngine.
Aside from being unnecessary to to a new QQmlEngine on the stack,
it makes the QML profiler not work.  Creating a new QQmlEngine consumes
the tcp connection used to communicate with the profiler, and as the
new engine is destroyed at the end of the scope, the profiler is
immediately disconnected.

Fix that by retrieving the already-constructed QQmlEngine using
the qmlEngine method.