Add "values" property lookup to StyleSet
This lookup function supports multiple values (like get() does), but
it will evaluate CSS expressions, such that property bindings like this
are possible:

  colors: rgb(100, 200, 50), rgb(50, 60, 70)

The "values" function should only be available in Aqt.StyleSheets
version 1.2, but since QML does not respect Q_REVISION on Q_INVOKABLES
currently (cf. QTBUG-30537) the introduction of protocol 2 is for
"correctness" only.  For the same reason we can't deprecate or rename
the get() lookup.