Add an experiment for using qdoc to generate documentation.
QDoc has the advantage that it already has all of the documentation
markup and code highlighting for QML APIs, but the disadvantage that
those are not extracted from header files, and require explicit
(separate) documentation of each signature of interest.

So, we would need to migrate the docs to some extent to use this.