Experiment with using Sphinx/Breathe to document QML API.
This has the advantage that the documentation can stay in the header,
and there is some control over the output and it is quite minimal.

Breathe is a good start, but it trips on Q_INVOKABLE and there are
some other rendering issues that would need to be resolved.

It also renders the class with a namespace, and I didn't yet try to
make it not do that.

What is really needed is a tool to generate documentation for one
language using an implementation in a different language. Such a general
tool doesn't seem to exist yet, but Breathe is nearly there.  I suspect
it would be possible to hack or monkey patch it to do what we want.