Remove parameter from signal StyleSetAttached::pathChanged
pathToString is potentially expensive. It is expected to be a rare case in
production code that QML clients listen to path changes *and* need to know
the new path. For this reason, building the path string whenever the
StyleSet.name property changes or an Item is reparented, seems to be unnecessary
work. Listeners to pathChanged could still query the path property afterwards.