Only connect to StyleEngineHost::styleEngineLoaded if necessary
If there is a StyleEngine at construction time, use it. Otherwise connect to
styleEngineLoaded to set it later and disconnect after setting it. In other
words, be prepared that the StyleEngine does not yet exist at construction time,
but once it's set, don't ever change it again. This underlines the current code
that ignores subsequent StyleEngine instances and removes the complexity of
dealing with StyleEngine changes. Also, it reduces the number of connections to
the styleEngineLoaded signal that would be unnecessary in the majority of cases.