Avoid crash in V4 in case the QML engine is recreated
The singleton StyleSetProps created by StyleSetProps::nullStyleSetProps() used
to survive QQmlEngine destruction. Accessing the singleton from a QObject
derived type in the context of a different engine led to assertion
"value.isString()" in the V4 runtime, eventually crashing.

Fix the crash by connecting the lifetime of the singleton to the lifetime of the
engine.

Make nullStyleSetProps a file local function since the new implementation is
quite specific to the StyleSet client (which is the only client anyway).