Fully include StyleSetProps in StyleEngine
Depending on the STL implementation the complete type StyleSetProps needs to be
available as it is used in a unique_ptr which is the value type of an
unordered_map. It looks like that this has worked only by coincidence. The
following refactorings change the includes and compile units in a way that would
lead to a compile error with Microsoft's STL.