Propagate parent changes down the hierarchy
This implementation requires the reparented item to have a StyleSet attached.
Otherwise the path of the descendants won't be updated.

An alternative would be to listen to the parentChanged signals of *all*
ancestors, including reconnection on hierarchy changes, but this obviously does
not scale which has also been proven by benchmarks. An optimization has been
tried to only listen to the parentChanged signals of the ancestors up to an
ancestor that has a StyleSet attached. Then only listen to some ancestorChanged
signal of that StyleSet which would result in a notification chain that would
skip a substantial amount of nodes in the parent hierarchy. But even with this
optimization the implementation has been shown to not scale well. Especially the
performance during destruction time has been observed to be bad because
QQuickItems emit the parentChanged signal from their destructor.

The hierarchy is traversed pre-order to ensure that the path of the StyleSet of
descendant A is up to date before attempting to refresh the path of the StyleSet
of object B which happens to be a descendant of A. This must be ensured because
B's StyleSet will reuse the path of A's StyleSet.