Prefer visual parent over QObject parent
Window types need a special case because they put their content in a root item
such that the resulting path would be e.g.
`QQuickRootItem/QQuickLoader/QQuickItem/QQuickItem/ContentItem` instead of
`ApplicationWindow`.

This commit also unifies the paths of ListView and Flickable. It would be
desirable to special case them as well to get rid of the unintuitive QQuickItem
in the path, but this is left for another branch.