Improve CMake check for VS compiler
Being on a Windows OS does not automatically mean that Visual Studio
compiler is used. If the compiler is MinGW's gcc for example the check
for the Visual Studio compiler version wouldn't have allowed this. This
is now fixed by checking for `MSVC` instead of `WIN32` before checking
the actual version of the compiler